Fix: format/engine detection in render_file()
The function had some issues detecting the file format if it contained
additional parts separated by dots. The file score.dom.js, for example
was not detected as javascript file.

Besides, the implementation was a mess, cleaned it up a bit. But I fear
this needs even more work to make it *maintainable*.